created a follow trajectory command that can follow any path from pp
