+++ simgui.json
@@ -63,6 +63,12 @@
     }
   },
   "NetworkTables": {
+    "LeftOrigin": {
+      "open": true
+    },
+    "LeftPiece": {
+      "open": true
+    },
     "Shuffleboard": {
       "Auton Configuration": {
         "open": false
